prepped a simple catches histogrammer that outputs something we can compare against Peter's data. Now I'd like to try an optimization run to test the concept further

+++ src/main/java/uk/ac/ox/oxfish/biology/complicated/StructuredAbundance.java
@@ -115,9 +115,9 @@ public class StructuredAbundance {
     }
 
 
-    public int getAbundanceInBin(int bin)
+    public double getAbundanceInBin(int bin)
     {
-        int fish = 0;
+        double fish = 0;
         for(int group = 0; group < getSubdivisions(); group++)
             fish += abundance[group][bin];
         return fish;
